frontend/dma: revert fifo_depth to 512 to ensure we can store a full SDCard block in FIFO.
When reducing fifo_depth, we make some assumptions on the available DMA speed in the SoC
(reads on the main bus should be faster than SDCard writes, writes on the main bus should be than
SDCard reads). This is true in most of the cases, but in some (like VexRiscv-SMP, this is not the
case. So for now use fifo_depth of 512 that does not require this assumption (but use a little more
resources).