Gen: Add initial/minimal standalone core generator.
Takes a clk/rst as inputs (with configurable frequency), connect to the SDCard pads
and provide a Wishbone slave interface for the control of the Core and a Wishbone
master interface for the DMAs to/from an external Memory.

The clk_frequency and FPGA vendor can be configured with command lines arguments.