core/phy: Add ability to wait while card signals that it is busy
SD cards can signal a busy condition for some commands (for example,
STOP_TRANSMISSION and ERASE) by pulling the DAT0 line low.  For those
commands, we need to wait, with the clock still running, while the
card is signalling that it is busy, and not signal command completion
until DAT0 goes high.

This adds a bit to the CMDCMD CSR, bit 2, so that software can indicate
that the command is one for which we need to wait while the card is
busy.  If that bit is set for a command, the SDPHYCMDR module will
send an extra byte via cmdr.source after the response, once the card
is not busy or a 1 second timeout has expired.  The SDCore module waits
for this extra byte before signalling command completion.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>