phy: Ensure we don't generate short pulses on the SD card clock
This latches the sdpads.clk signal when the SD card clock goes high
so that we don't generate short pulses on the SD card clock (i.e.
where the clock goes high for a shorter time than the normal clock
high time for the selected clock rate).

This is done by latching the sdpads.clk enable signal at the point
where the SD card clock goes high (i.e. the last system clock cycle
where clk_d is high) and then using the latched enable for the rest
of the time when the SD card clock is high.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>